Undo the duplicate arena info and apt price fetches additions from the layout
