+++ src/typescript/frontend/src/app/wallet/[address]/page.tsx
@@ -2,6 +2,8 @@
 
 import { sha3_256 } from "@noble/hashes/sha3";
 import { WalletClientPage } from "components/pages/wallet/WalletClientPage";
+import { AptPriceContextProvider } from "context/AptPrice";
+import { fetchCachedAptPrice } from "lib/queries/get-apt-price";
 import generateMetadataHelper from "lib/utils/generate-metadata-helper";
 import type { Metadata } from "next";
 
@@ -49,9 +51,13 @@ export default async function WalletPage({ params }: Props) {
     return <UserNotFound />;
   }
 
+  const aptPrice = await fetchCachedAptPrice();
+
   return (
     <div className="mx-auto">
-      <WalletClientPage address={address} name={name} />
+      <AptPriceContextProvider aptPrice={aptPrice}>
+        <WalletClientPage address={address} name={name} />
+      </AptPriceContextProvider>
     </div>
   );
 }
